Separate MainController into it's own file
